JAVA-41270 Moved code of Obtaining Auto-generated Keys in Spring JDBC to spring-jdbc-2
