+++ persistence-modules/spring-jdbc-2/README.md
@@ -2,3 +2,4 @@
 - [A Guide to Spring 6 JdbcClient API](https://www.baeldung.com/spring-6-jdbcclient-api)
 - [Connect with PostgreSQL Database over SSL](https://www.baeldung.com/spring-boot-jdbc-postgresql-ssl)
 - [Stored Procedures With Spring JdbcTemplate](https://www.baeldung.com/spring-jdbctemplate-stored-procedure)
+- [Obtaining Auto-generated Keys in Spring JDBC](https://www.baeldung.com/spring-jdbc-autogenerated-keys)
