+++ persistence-modules/spring-jdbc-2/src/main/resources/com/baeldung/spring/jdbc/autogenkey/autogenkey-schema.sql
@@ -1,0 +1,5 @@
+CREATE TABLE IF NOT EXISTS sys_message (
+  id bigint  NOT NULL AUTO_INCREMENT,
+  message varchar(100) NOT NULL,
+  PRIMARY KEY (id)
+);
