--- persistence-modules/spring-jdbc/src/main/resources/com/baeldung/spring/jdbc/autogenkey/autogenkey-schema.sql
@@ -1,5 +1,0 @@
-CREATE TABLE IF NOT EXISTS sys_message (
-  id bigint  NOT NULL AUTO_INCREMENT,
-  message varchar(100) NOT NULL,
-  PRIMARY KEY (id)
-);
